third_party/WebKit/LayoutTests/fast/mediastream/getMediaDevices-expected.txt

Tests navigator.getMediaDevices

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.getMediaDevices(devicesCallback); did not throw exception.
PASS After getMediaDevices
PASS Devices returned.
PASS devices.length is 3
PASS devices[0].deviceId.length is > 0
PASS devices[1].deviceId.length is > 0
PASS devices[2].deviceId.length is > 0
PASS devices[0].deviceId is not devices[1].deviceId
PASS devices[0].deviceId is not devices[2].deviceId
PASS devices[1].deviceId is not devices[2].deviceId
PASS devices[0].kind is "audioinput"
PASS devices[1].kind is "audiooutput"
PASS devices[2].kind is "videoinput"
PASS devices[0].label.length is > 0
PASS devices[1].label.length is > 0
PASS devices[2].label.length is > 0
PASS devices[0].label is not devices[1].label
PASS devices[0].label is not devices[2].label
PASS devices[1].label is not devices[2].label
PASS devices[0].groupId.length is > 0
PASS devices[1].groupId.length is > 0
PASS devices[2].groupId.length is > 0
PASS devices[0].groupId is devices[1].groupId
PASS devices[0].groupId is not devices[2].groupId
PASS successfullyParsed is true

TEST COMPLETE
